requirements.txt
PY_GUI
pygame
python-dotenv
cryptography
requests
pillow